+++ src/seek/mod.rs
@@ -4,4 +4,6 @@ mod search;
 
 pub use scan::scan;
 pub use scan_result::ScanResult;
+pub use search::filter_excluded_dirs;
+pub use search::filter_included_dirs;
 pub use search::search;